Fix clang build warnings and errors on Mac (DB-502 and DB-876).
- Remove unused function is_null_field in ha_tokudb.cc.
- Use portable printf format specifier for uint64_t in
  ha_tokudb::analyze.
- Avoid "[enum] val always compares to less than 256" warning by
  casting the enum to int in
  hatoku_cmp.cc:create_toku_key_descriptor_for_key.
- Adjust tokudb_hton_name to be const char * consistently everywhere.
- Fix tokudb_hcad_update_fun returning uninitilised error in case of
  OOM. Return ENOMEM instead.

+++ storage/tokudb/hatoku_cmp.cc
@@ -1026,7 +1026,7 @@ static int create_toku_key_descriptor_for_key(KEY* key, uchar* buf) {
         // The second byte for each field is the type
         //
         TOKU_TYPE type = mysql_to_toku_type(field);
-        assert (type < 256);
+        assert ((int)type < 256);
         *pos = (uchar)(type & 255);
         pos++;
